# feat: Implement Rust DataFusion Engine for 10x Performance Improvement