--- a/bigquery-lite-engine/Cargo.toml
+++ b/bigquery-lite-engine/Cargo.toml
@@ -52,6 +52,7 @@ default = ["object_store"]
 [dev-dependencies]
 tempfile = "3.8"
 criterion = { version = "0.5", features = ["html_reports"] }
+tokio-test = "0.4"
 
 [[bench]]
 name = "query_benchmark"