… backend/tests/test_error_handling.py
… backend/tests/test_memory_validation.py
… backend/tests/test_performance_benchmarks.py
… backend/tests/test_rust_integration.py
… bigquery-lite-engine/Cargo.lock
… bigquery-lite-engine/src/benchmarks.rs
… bigquery-lite-engine/src/utils.rs
… bigquery-lite-engine/tests/engine_tests.rs
… run_all_tests.py
… validate_rust_engine.py